Added Town of Salem
That's the best way to define it, added most of the roles and their interactions. Added unit tests. Added visiting, attack and defense mechanics. Added priority mechanics. Fixed typos in role names. Fixed name of Escort and Consort to reflect the new names Tavern Keeper and Bootlegger. Added many many flags to Player(), these flags are crucial to determining deaths as well as win conditions. For coven roles in which the priority changes based on the possession of the Necronomicon, this should also be done.